Create common DiffViewerDialog & use in code block
